+++ browser/js/twitch/content/web/community_points_auto_claim.js
@@ -1,0 +1,18 @@
+'use strict';
+
+import { observeSearchCommunityPointsSummary,
+         observeGetCommunityPointsSummaryClaimButton } from './twitch_objects.mjs';
+
+function processClaimButton(claimButton){
+    claimButton.click();
+}
+
+function processCommunityPointsSummary(communityPointsSummary){
+    observeGetCommunityPointsSummaryClaimButton(processClaimButton, communityPointsSummary);
+}
+
+function setup(){
+    observeSearchCommunityPointsSummary(processCommunityPointsSummary);
+}
+
+setup();